chore: rename legacy fighter portrait
