+++ frontend/src/lib/battle/LegacyFighterPortrait.svelte
